Decouple share and schedule buttons from parent components
