+++ projects/js-packages/publicize-components/src/components/social-post-modal/settings-section.tsx
@@ -2,12 +2,7 @@ import { Panel } from '@wordpress/components';
 import { useSelect } from '@wordpress/data';
 import { store as editorStore } from '@wordpress/editor';
 import { __, _x } from '@wordpress/i18n';
-import { useCallback } from 'react';
-import { useIsReSharingPossible } from '../../hooks/use-is-resharing-possible';
 import usePublicizeConfig from '../../hooks/use-publicize-config';
-import { useSchedulePost } from '../../hooks/use-schedule-post';
-import useSocialMediaConnections from '../../hooks/use-social-media-connections';
-import { store as socialStore } from '../../social-store';
 import { SharePostForm } from '../form/share-post-form';
 import ScheduleButton from '../schedule-button';
 import { SharePostButton } from '../share-post';
@@ -23,29 +18,9 @@ import styles from './styles.module.scss';
  */
 export function SettingsSection( { onReShared } ) {
 	const isPostPublished = useSelect( select => select( editorStore ).isCurrentPostPublished(), [] );
-	const isSavingPost = useSelect( select => select( editorStore ).isSavingPost(), [] );
 	const postId = useSelect( select => select( editorStore ).getCurrentPostId(), [] );
 
 	const { isRePublicizeUpgradableViaUpsell } = usePublicizeConfig();
-	const isReSharingPossible = useIsReSharingPossible();
-	const { enabledConnections } = useSocialMediaConnections();
-
-	const { schedulePost } = useSchedulePost();
-
-	const isSavingScheduledShare = useSelect(
-		select => select( socialStore ).isSavingScheduledShare(),
-		[]
-	);
-
-	const onSchedule = useCallback(
-		async scheduleTimestamp => {
-			await schedulePost( {
-				connectionIds: enabledConnections.map( connection => Number( connection.connection_id ) ),
-				timestamp: scheduleTimestamp,
-			} );
-		},
-		[ schedulePost, enabledConnections ]
-	);
 
 	return (
 		<div className={ styles[ 'settings-section' ] }>
@@ -66,15 +41,8 @@ export function SettingsSection( { onReShared } ) {
 				<SharePostForm analyticsData={ { location: 'preview-modal' } } />
 				{ isPostPublished && ! isRePublicizeUpgradableViaUpsell && (
 					<div className={ styles[ 'share-actions' ] }>
-						<ScheduleButton
-							isDisabled={ ! isReSharingPossible }
-							onConfirm={ onSchedule }
-							isBusy={ isSavingScheduledShare || isSavingPost }
-						/>
-						<SharePostButton
-							onShareCompleted={ onReShared }
-							isDisabled={ isSavingScheduledShare }
-						/>
+						<ScheduleButton />
+						<SharePostButton onShareCompleted={ onReShared } />
 					</div>
 				) }
 			</div>
